[IMP] mail: return focus back to thread after using quick reactions
Previously, closing the quick reaction menu did not return focus to
the thread, which was inconvenient—especially in chat windows where
pressing Escape is commonly used to close them.

This update ensures that focus is restored to the message composer if
one is being edited; otherwise, it returns to the thread composer.
This improves usability, allowing users to react to messages without
disrupting their workflow.

closes odoo/odoo#198261

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>